hw/mos6522: Simplify get_next_irq_time()
The time at which the next timer interrupt will occur (in nanoseconds)
is derived from the time at which the timer was loaded (also in
nanoseconds). Part of the calculation is performed in units of cycles
so unit conversions are required. However, there is no requirement to
use 16-bit values (i.e. the hardware representation). It just complicates
things.