hw/mos6522: Synchronize timer interrupt and timer counter
We rely on a QEMUTimer callback to set the interrupt flag, but this
appears to race with counter register accesses, such that the guest
might see the counter reloaded, but not see the interrupt flagged yet.

According to the datasheet, a real 6522 device counts down to FFFF, then
raises the relevant IRQ. After the FFFF count, the counter reloads from
the latch (for timer 1) or continues to decrement thru FFFE (for timer 2).

Therefore, the guest operating system may read zero from T1CH and infer
that the counter has not yet wrapped (given the interrupt flag isn't
set).

Similarly, the guest may find the timer interrupt flag to be set and
infer that the counter is non-zero (given another full count hasn't yet
elapsed).

Synchronize the timer counter and interrupt flag such that the guest will
always observe the correct sequence of states.